refactored user manager
